Refractored login code from HomeController to LoginController
